+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 * Unreleased
+    * Add `copyReplaceChar()` and `copyReplaceString()` functions. Extracted
+      from AceTime library.
 * 1.4.3 (2021-02-18)
     * Add `linearSearch()` and `linearSearchByKey()`. Analogs to the
       corresponding binary search functions.